feat(structure): create useDocumentPaneInitialValue
